third_party/blink/web_tests/fast/forms/search/input-search-press-escape-key-expected.txt

PASS input.value is ""
PASS searchEventObserved is true
PASS input.value is ""
PASS searchEventObserved is false
PASS input.value is "foo"
PASS searchEventObserved is false
PASS input.value is "foo"
PASS searchEventObserved is false
PASS input.value is "foo"
PASS searchEventObserved is false
PASS successfullyParsed is true

TEST COMPLETE


This tests if the value in a search input form is cleared and a 'search' event is triggered, when we press the Escape key. To run (a part of) this test manually, type some text in the search form and then press the Escape key. If the text is cleared, then the test passes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
